+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.0.38
+
+- Hello, Menu Bar! sample
+- Support `menu` and `menu_item` including `:separator`, `:radiobutton`, and `:checkbutton` styles
+- Support menu item accelerators
+
 ## 0.0.37
 
 - Support `focus: true` or `focus: [integer]` option for all selection formatting methods (`add_selection_format`, `remove_selection_format`, `toggle_selection_format`, `add_selection_font_format`, `remove_selection_font_format`, `toggle_selection_font_format`)